requirements.txt
# ============================================
# REQUIREMENTS.TXT OPTIMIZADO PARA RAILWAY
# ============================================

# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Autenticación
Flask-Login==0.6.3

# Base de datos
sqlite3

# Email
secure-smtplib==0.1.1

# PDF Generation
reportlab==4.0.7

# Variables de entorno
python-dotenv==1.0.0

# Seguridad
MarkupSafe==2.1.3

# Compresión (opcional)
Flask-Compress==1.14

# Producción
gunicorn==21.2.0

# Optimización
Pillow==10.1.0

# ============================================
# NOTAS DE INSTALACIÓN
# ============================================
# 
# Railway instalará automáticamente estas dependencias
# Si hay problemas, verificar que todas las versiones sean compatibles
# 
# Para desarrollo local:
# pip install -r requirements.txt
# 
# Para Railway:
# Las dependencias se instalan automáticamente al hacer deploy